requirements.txt
pymysql
python-dotenv
openai